+++ svc-artisans/src/main/java/com/arquiproject/svc_artisans/service/IReviewService.java
@@ -4,7 +4,7 @@ import com.arquiproject.svc_artisans.model.Review;
 
 public interface IReviewService {
     Review createReview(Review review);
-    boolean deleteReview (int reviewId);
+    boolean deleteReview (Long reviewId);
     Review updateReview(Review review);
 
 }
